# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Ads Platform — merchant dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MeResponse {
  merchantId: string;
  storeName: string;
  ownerName?: string | null;
  plan: string;
  status: string;
  consentAccepted: boolean;
  connectedPlatforms: string[];
}

export interface ConsentBody {
  readStoreData: boolean;
  receiveWebhooks: boolean;
  shareAudienceNetwork: boolean;
  manageAdAccounts: boolean;
}

export type AudienceSizeTopCitiesItem = {
  city: string;
  share: number;
};

export interface AudienceSize {
  segmentName?: string;
  segmentDisplayName: string;
  totalBuyers: number;
  totalMerchants: number;
  topCities: AudienceSizeTopCitiesItem[];
}

export interface AdAccountSummary {
  id: string;
  platform: string;
  accountIdExternal: string;
  status: string;
  connectedAt: string;
}

export interface ConnectUrl {
  url: string;
  available: boolean;
  reasonArabic?: string | null;
}

export interface Greeting {
  greetingArabic: string;
  subtitleArabic: string;
  merchantName: string;
}

export type TodayMetricsDeltaVsYesterday = {
  ordersPct: number;
  spendPct: number;
  revenuePct: number;
};

export interface TodayMetrics {
  orders: number;
  spend: number;
  revenue: number;
  costPerOrder?: number | null;
  roas?: number | null;
  currency: string;
  deltaVsYesterday: TodayMetricsDeltaVsYesterday;
}

export interface AiDecisionItem {
  id: string;
  decisionType: string;
  reasonArabic: string;
  resultArabic?: string | null;
  executedAt: string;
}

export interface PlatformPerformance {
  platform: string;
  platformLabelArabic: string;
  orders: number;
  spend: number;
  costPerOrder?: number | null;
  roas?: number | null;
  share: number;
}

export type CityBreakdownTopDistrictsItem = {
  district: string;
  orders: number;
};

export interface CityBreakdown {
  city: string;
  orders: number;
  revenue: number;
  topDistricts: CityBreakdownTopDistrictsItem[];
}

export interface SeasonalAlertItem {
  id: string;
  nameArabic: string;
  triggerDate: string;
  daysUntil: number;
  targetCategories: string[];
}

export type GetAdAccountConnectUrlParams = {
  platform: GetAdAccountConnectUrlPlatform;
};

export type GetAdAccountConnectUrlPlatform =
  (typeof GetAdAccountConnectUrlPlatform)[keyof typeof GetAdAccountConnectUrlPlatform];

export const GetAdAccountConnectUrlPlatform = {
  meta: "meta",
  snap: "snap",
  tiktok: "tiktok",
  google: "google",
} as const;

export type ListAiDecisionsParams = {
  limit?: number;
};