# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Ads Platform — merchant dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/consent:
    post:
      operationId: saveConsent
      tags: [onboarding]
      summary: Persist merchant consent toggles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentBody"
      responses:
        "200":
          description: Consent saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
...
  /ad-accounts/connect-url:
    get:
      operationId: getAdAccountConnectUrl
      tags: [ad-accounts]
      summary: Returns the platform's OAuth URL to start connection
      parameters:
        - in: query
          name: platform
          required: true
          schema:
            type: string
            enum: [meta, snap, tiktok, google]
      responses:
        "200":
          description: OAuth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectUrl"
  /ad-accounts/{platform}/mock-connect:
    post:
      operationId: mockConnectAdAccount
      tags: [ad-accounts]
      summary: Dev-only — pretend that the given platform has been connected so the merchant can keep onboarding before real OAuth credentials exist.
      parameters:
        - in: path
          name: platform
          required: true
          schema:
            type: string
            enum: [meta, snap, tiktok, google]
      responses:
        "200":
          description: Mock account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccountSummary"
...
components:
  schemas:
...
    MeResponse:
      type: object
      properties:
        merchantId:
          type: string
        storeName:
          type: string
        ownerName:
          type: string
          nullable: true
        plan:
          type: string
        status:
          type: string
        consentAccepted:
          type: boolean
        connectedPlatforms:
          type: array
          items:
            type: string
      required:
        - merchantId
        - storeName
        - plan
        - status
        - consentAccepted
        - connectedPlatforms
    ConsentBody:
      type: object
      properties:
        readStoreData:
          type: boolean
        receiveWebhooks:
          type: boolean
        shareAudienceNetwork:
          type: boolean
        manageAdAccounts:
          type: boolean
      required:
        - readStoreData
        - receiveWebhooks
        - shareAudienceNetwork
        - manageAdAccounts
...
    AdAccountSummary:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
        accountIdExternal:
          type: string
        status:
          type: string
        connectedAt:
          type: string
          format: date-time
      required:
        - id
        - platform
        - accountIdExternal
        - status
        - connectedAt
    ConnectUrl:
      type: object
      properties:
        url:
          type: string
        available:
          type: boolean
        reasonArabic:
          type: string
          nullable: true
      required: [url, available]
...