# lib/api-zod/src/generated/types/consentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Ads Platform — merchant dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentBody {
  readStoreData: boolean;
  receiveWebhooks: boolean;
  shareAudienceNetwork: boolean;
  manageAdAccounts: boolean;
}