# lib/api-zod/src/generated/types/meResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Ads Platform — merchant dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface MeResponse {
  merchantId: string;
  storeName: string;
  ownerName?: string | null;
  plan: string;
  status: string;
  consentAccepted: boolean;
  connectedPlatforms: string[];
}